requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
watchdog==3.0.0
schedule==1.2.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml
pandas
opendatasets==0.1.22
SpeechRecognition==3.10.1
pydub==0.25.1
flask-caching==2.1.0
flask-cors==4.0.0
google-generativeai==0.3.1
